# backend/src/test/resources/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: 공통기능 25개 API
  version: 0.1.0
servers:
- url: /
paths:
...
  /api/admin/menus:
    get:
      operationId: listMenus
      summary: 메뉴 관리 조회
      x-related-requirements:
      - REQ-032
      - REQ-033
      - REQ-034
      - REQ-035
      - REQ-105
      - REQ-107
      - REQ-114
      - REQ-115
      - REQ-116
      - REQ-117
      - REQ-119
      - REQ-122
      - REQ-123
      - REQ-124
      - REQ-126
      - REQ-145
      - REQ-166
      - REQ-167
      - REQ-177
      - REQ-194
      - REQ-198
      - REQ-218
      x-roles:
      - R09
      parameters:
      - name: page
        in: query
        schema:
          type: integer
          default: 1
      - name: size
        in: query
        schema:
          type: integer
          enum:
          - 20
          - 50
          - 100
          default: 20
      - name: sort
        in: query
        schema:
          type: string
      - name: q
        in: query
        schema:
          type: string
      - name: filter
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
    post:
      operationId: saveMenus
      summary: 메뉴 관리 저장/실행
      x-related-requirements:
      - REQ-032
      - REQ-033
      - REQ-034
      - REQ-035
      - REQ-105
      - REQ-107
      - REQ-114
      - REQ-115
      - REQ-116
      - REQ-117
      - REQ-119
      - REQ-120
      - REQ-122
      - REQ-126
      - REQ-133
      - REQ-134
      - REQ-145
      - REQ-166
      - REQ-167
      - REQ-177
      - REQ-194
      - REQ-198
      - REQ-218
      x-roles:
      - R09
      x-business-rules:
      - 메뉴 관리 원문 제약과 서버 권한을 적용한다
      - menus 저장 전 필수값·기간·중복 조건을 검증한다
      x-side-effects:
      - menus 레코드를 등록·수정하거나 실행 이력을 기록한다
      - audit_logs에 메뉴 관리 변경 전후 값을 기록한다
      x-required-tests:
      - 'happy: 유효한 메뉴 관리 저장 -> menus 후속 조회에 반영'
      - 'auth: R01이 메뉴 관리 저장 요청 -> 403 반환'
      - 'validation: 필수 식별자 누락 -> fields.id 오류'
      - 'business: 원문 불변/기간/중복 제약 위반 -> menus 상태 변경 없음'
      - 'side-effect: 성공 처리 후 audit_logs에 메뉴 관리 이력 존재'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
...
  /api/admin/excel-templates:
    get:
      operationId: listExcelTemplates
      summary: 업로드 양식 관리 조회
      x-related-requirements:
      - REQ-063
      - REQ-064
      - REQ-065
      - REQ-105
      - REQ-110
      - REQ-114
      - REQ-115
      - REQ-116
      - REQ-117
      - REQ-123
      - REQ-124
      - REQ-126
      - REQ-153
      - REQ-166
      - REQ-167
      - REQ-185
      - REQ-194
      - REQ-198
      - REQ-210
      - REQ-218
      x-roles:
      - R09
      parameters:
      - name: page
        in: query
        schema:
          type: integer
          default: 1
      - name: size
        in: query
        schema:
          type: integer
          enum:
          - 20
          - 50
          - 100
          default: 20
      - name: sort
        in: query
        schema:
          type: string
      - name: q
        in: query
        schema:
          type: string
      - name: filter
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
    post:
      operationId: saveExcelTemplates
      summary: 업로드 양식 관리 저장/실행
      x-related-requirements:
      - REQ-063
      - REQ-064
      - REQ-065
      - REQ-105
      - REQ-110
      - REQ-114
      - REQ-115
      - REQ-116
      - REQ-117
      - REQ-120
      - REQ-126
      - REQ-133
      - REQ-134
      - REQ-153
      - REQ-166
      - REQ-167
      - REQ-185
      - REQ-194
      - REQ-198
      - REQ-210
      - REQ-218
      x-roles:
      - R09
      x-business-rules:
      - 업로드 양식 관리 원문 제약과 서버 권한을 적용한다
      - excel_templates 저장 전 필수값·기간·중복 조건을 검증한다
      x-side-effects:
      - excel_templates 레코드를 등록·수정하거나 실행 이력을 기록한다
      - audit_logs에 업로드 양식 관리 변경 전후 값을 기록한다
      x-required-tests:
      - 'happy: 유효한 업로드 양식 관리 저장 -> excel_templates 후속 조회에 반영'
      - 'auth: R01이 업로드 양식 관리 저장 요청 -> 403 반환'
      - 'validation: 필수 식별자 누락 -> fields.id 오류'
      - 'business: 원문 불변/기간/중복 제약 위반 -> excel_templates 상태 변경 없음'
      - 'side-effect: 성공 처리 후 audit_logs에 업로드 양식 관리 이력 존재'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
  /api/admin/excel-uploads:
    get:
      operationId: uploadExcel
      summary: 엑셀 업로드 조회
      x-related-requirements:
      - REQ-066
      - REQ-067
      - REQ-068
      - REQ-069
      - REQ-070
      - REQ-071
      - REQ-105
      - REQ-110
      - REQ-114
      - REQ-115
      - REQ-116
      - REQ-117
      - REQ-120
      - REQ-123
      - REQ-124
      - REQ-126
      - REQ-133
      - REQ-134
      - REQ-154
      - REQ-166
      - REQ-167
      - REQ-186
      - REQ-194
      - REQ-198
      - REQ-210
      x-roles:
      - R09
      parameters:
      - name: page
        in: query
        schema:
          type: integer
          default: 1
      - name: size
        in: query
        schema:
          type: integer
          enum:
          - 20
          - 50
          - 100
          default: 20
      - name: sort
        in: query
        schema:
          type: string
      - name: q
        in: query
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
    post:
      operationId: uploadExcel
      summary: 엑셀 업로드 저장/실행
      x-related-requirements:
      - REQ-066
      - REQ-067
      - REQ-068
      - REQ-069
      - REQ-070
      - REQ-071
      - REQ-105
      - REQ-110
      - REQ-114
      - REQ-115
      - REQ-116
      - REQ-117
      - REQ-120
      - REQ-123
      - REQ-124
      - REQ-126
      - REQ-133
      - REQ-134
      - REQ-154
      - REQ-166
      - REQ-167
      - REQ-186
      - REQ-194
      - REQ-198
      - REQ-210
      x-roles:
      - R09
      x-business-rules:
      - 엑셀 업로드 원문 제약과 서버 권한을 적용한다
      - excel_upload_histories 저장 전 필수값·기간·중복 조건을 검증한다
      x-side-effects:
      - excel_upload_histories 레코드를 등록·수정하거나 실행 이력을 기록한다
      - audit_logs에 엑셀 업로드 변경 전후 값을 기록한다
      x-required-tests:
      - 'happy: 유효한 엑셀 업로드 저장 -> excel_upload_histories 후속 조회에 반영'
      - 'auth: R01이 엑셀 업로드 저장 요청 -> 403 반환'
      - 'validation: 필수 식별자 누락 -> fields.id 오류'
      - 'business: 원문 불변/기간/중복 제약 위반 -> excel_upload_histories 상태 변경 없음'
      - 'side-effect: 성공 처리 후 audit_logs에 엑셀 업로드 이력 존재'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
...
  /api/admin/excel-uploads/{uploadId}/errors/download:
    get:
      operationId: downloadExcelUploadErrors
      summary: 엑셀 오류목록 다운로드
      x-related-requirements:
      - REQ-066
      - REQ-067
      - REQ-068
      - REQ-069
      - REQ-070
      - REQ-071
      - REQ-110
      - REQ-114
      - REQ-115
      - REQ-116
      - REQ-117
      - REQ-126
      - REQ-154
      - REQ-166
      - REQ-167
      - REQ-186
      - REQ-198
      - REQ-210
      x-roles:
      - R09
      x-business-rules:
      - 엑셀 오류목록 다운로드 권한과 원문 제약을 검증한다
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
  /api/admin/excel-templates/{templateId}/download:
    get:
      operationId: downloadExcelTemplate
      summary: 업로드 양식 다운로드
      x-related-requirements:
      - REQ-063
      - REQ-064
      - REQ-065
      - REQ-110
      - REQ-114
      - REQ-115
      - REQ-116
      - REQ-117
      - REQ-126
      - REQ-153
      - REQ-166
      - REQ-167
      - REQ-185
      - REQ-198
      - REQ-210
      x-roles:
      - R09
      x-business-rules:
      - 업로드 양식 다운로드 권한과 원문 제약을 검증한다
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: SESSION
  schemas:
    ApiResponse:
      type: object
      required:
      - success
      - meta
      properties:
        success:
          type: boolean
          enum:
          - true
        data: {}
        meta:
          type: object
    ApiError:
      type: object
      required:
      - success
      - meta
      properties:
        success:
          type: boolean
          enum:
          - false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            fields:
              type: object
        meta:
          type: object
...
security:
- sessionCookie: []
...